requirements.txt
qiskit>=1.0.0
qiskit-aer>=0.13.0
qiskit-ibm-runtime>=0.20.0
numpy>=1.24.0,<2.0.0
scipy>=1.10.0
pandas>=2.0.0
matplotlib>=3.7.0
seaborn>=0.12.0
pyyaml>=6.0